refactor: domain 디렉토리 생성
