+++ src/domain/class/__test__/CarInfo.test.js
