+++ src/domain/class/__test__/RaceManager.test.js
